new seeds, updated db

+++ db/migrate/20200902174921_create_questions.rb
@@ -3,8 +3,8 @@ class CreateQuestions < ActiveRecord::Migration[6.0]
     create_table :questions do |t|
       t.integer :game_id
       t.string :question
-      t.integer :answer
-      t.integer :over_under_line
+      t.float :answer
+      t.float :over_under_line
       t.string :img_url
 
       t.timestamps
